refactor(private-world): relocate core modules
