+++ tests/test_packaging_local_server.py
@@ -80,6 +80,7 @@ def test_wheel_installs_every_module_needed_to_import_local_server(
     assert "runtime/persona/persona_loader.py" in packaged_paths
     assert "runtime/persona/persona_assembly.py" in packaged_paths
     assert "runtime/persona/persona_provider.py" in packaged_paths
+    assert "runtime/private_world/ledger.py" in packaged_paths
     assert "reply_pipeline.py" in packaged_paths
     assert "reply_reviewer.py" in packaged_paths
     assert "reply_context.py" in packaged_paths
@@ -188,6 +189,11 @@ def test_wheel_installs_every_module_needed_to_import_local_server(
             "assert persona_loader is persona_loader_impl; "
             "assert persona_assembly is persona_assembly_impl; "
             "assert persona_provider is persona_provider_impl; "
+            "import private_world_port, private_world_commands, private_world_ledger; "
+            "import runtime.private_world.port, runtime.private_world.commands, runtime.private_world.ledger; "
+            "assert private_world_port is runtime.private_world.port; "
+            "assert private_world_commands is runtime.private_world.commands; "
+            "assert private_world_ledger is runtime.private_world.ledger; "
             "import music_reply, original_client_server, runtime.media.music_caption, song_content, voice_direction; "
             "import latentsync_reply; "
             "from runtime.media import latentsync_reply as canonical_latentsync_reply; "
